feat: add begin and end options
Added two options that enable the user to decide which parts of the
video to download. The begin function downloads the whole video till the
end, either the one specified with the --end option or the default one.
This is not optimal since ytdl-core has a begin option which lets you
specify the point of the video from which to start downloading, this
theoretically would enable us to download only the parts of the video
that are needed but I couldn't manage to get the ytdl-core's begin option
to work, so, for now, it'll work like this. The end option on the other
hand interrupts the download when the desired duration is achieved.
Plus I noticed that sometimes when using the begin and end option the
video doesn't download correctly, resulting in only the audio track
being present without the video. I'll investigate further and hopefully
find the source of the issue.