feat: add file input option
Added a file input option, if a file path is provided to the --l
parameter instead of a YouTube link, the program will loop through the
file line by line and download the YouTube videos located on those lines.
This allows the user to create a file with multiple YouTube links and
download them by passing the file as input to the program.
All other options like --mp3 and --thumbnail are also functional while
using a file input.