fix: improve path concatenation
Improved path concatenation, now the user can specify a filename in the
output option, before it was only possibly to specify an output folder
and the program automatically added a filename.
The program automatically detects if a user has supplied a folder path
or a complete path with filename, if it lacks the filename it'll add the
video title at the end, and if it was supplied then the output is
unaltered and only some normalization is done.